refactor: split engine.go/writes.go, rename rules2.go (#237)
Pure code-motion refactor, no behavior change:

- internal/capture/engine.go (2264 lines) split into engine.go (Engine
  struct, NewEngine/newEngineWith, Run, preflight, and a handful of small
  cross-cutting helpers), poll.go (pollResource/watchResource/streamWatch
  and the per-resource fetch loop), fetch.go (doFetch/rawFetch/
  fetchDiscovery*), demux.go (cluster-wide demux helpers), discover.go
  (autoDiscoverResources/nativeListableKinds/captureCoreTableSchemas/
  fetchTableSchema), podlogs.go (pod-log fetching), and record_writer.go
  (storeRecord).
- internal/server/writes.go (1783 lines) split into writes.go (the
  HTTP verb handlers), defaults.go (applyKnownDefaults/defaultObject/
  synthetic*IP), patch.go (applyPatch/strategicMergePatch/
  jsonMergeOrPatch), and scale.go (the /scale subresource handler and
  its helpers).
- internal/diagnose/rules2.go renamed to rules_workload.go (git mv,
  no content changes) — it was never a "v2" of rules.go, just
  misnamed.

internal/server/handler.go is intentionally left untouched: the issue
assumed the k9s SelfSubjectAccessReview shims and exec/portforward
rejection lived in separate helper methods, but they're actually
inline blocks inside ServeHTTP with no top-level declaration to
relocate — extracting them into new functions would be a structural
change, not just code motion, so out of scope here.

Every moved declaration keeps its exact body and doc comment; only its
file and each file's import block changed. gofmt, go vet, go build,
go test ./..., and go mod tidy all pass clean.
