+++ internal/server/writes.go
@@ -3,9 +3,7 @@ package server
 import (
 	"encoding/json"
 	"fmt"
-	"hash/fnv"
 	"io"
-	"net"
 	"net/http"
 	"sort"
 	"strconv"
@@ -13,19 +11,9 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	jsonpatch "gopkg.in/evanphx/json-patch.v4"
-	appsv1 "k8s.io/api/apps/v1"
-	batchv1 "k8s.io/api/batch/v1"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/apimachinery/pkg/util/strategicpatch"
 	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/utils/ptr"
 	"sigs.k8s.io/yaml"
 
 	kstore "github.com/phenixblue/k8shark/internal/store"
@@ -941,220 +929,6 @@ func (h *handler) objectForRead(group, version, resource, namespace, name string
 	return body, code == 200
 }
 
-// scaleSubresourceResources are the built-in Kinds with a real /scale
-// subresource on a live apiserver (Deployment, ReplicaSet, StatefulSet,
-// ReplicationController) — anything else 404s, matching upstream.
-var scaleSubresourceResources = map[string]bool{
-	"deployments":            true,
-	"replicasets":            true,
-	"statefulsets":           true,
-	"replicationcontrollers": true,
-}
-
-// scaleSelectorString renders a resource's .spec.selector as the label-query
-// string HPA (and kubectl) read from a Scale's status.selector — handling
-// both selector shapes real Kubernetes types use: Deployment/ReplicaSet/
-// StatefulSet's structured metav1.LabelSelector ({matchLabels,
-// matchExpressions}), and ReplicationController's plain map[string]string.
-// Trying the map shape first is required, not just sufficient: unmarshaling a
-// flat map like {"app":"foo"} into a LabelSelector struct silently succeeds
-// with a zero-value result (unknown JSON fields are ignored on structs), so
-// checking LabelSelector first would always win and produce an empty
-// selector for every ReplicationController. A real LabelSelector's fields
-// nest an object/array under "matchLabels"/"matchExpressions", which cannot
-// decode into a map[string]string value, so that shape reliably fails first.
-func scaleSelectorString(raw json.RawMessage) string {
-	if len(raw) == 0 {
-		return ""
-	}
-	var flat map[string]string
-	if err := json.Unmarshal(raw, &flat); err == nil {
-		return labels.SelectorFromSet(flat).String()
-	}
-	var sel metav1.LabelSelector
-	if err := json.Unmarshal(raw, &sel); err != nil {
-		return ""
-	}
-	selector, err := metav1.LabelSelectorAsSelector(&sel)
-	if err != nil {
-		return ""
-	}
-	return selector.String()
-}
-
-// scaleObject synthesizes an autoscaling/v1 Scale representation of a
-// Deployment/ReplicaSet/StatefulSet/ReplicationController — the real
-// apiserver does the same conversion server-side for its generic scale
-// subresource (e.g. pkg/registry/apps/deployment/storage's scaleClient).
-func scaleObject(namespace, name string, obj json.RawMessage) (json.RawMessage, error) {
-	var o struct {
-		Metadata struct {
-			UID             string `json:"uid"`
-			ResourceVersion string `json:"resourceVersion"`
-		} `json:"metadata"`
-		Spec struct {
-			Replicas *int32          `json:"replicas"`
-			Selector json.RawMessage `json:"selector"`
-		} `json:"spec"`
-		Status struct {
-			Replicas int32 `json:"replicas"`
-		} `json:"status"`
-	}
-	if err := json.Unmarshal(obj, &o); err != nil {
-		return nil, err
-	}
-	var replicas int32
-	if o.Spec.Replicas != nil {
-		replicas = *o.Spec.Replicas
-	}
-	// HPA (and kubectl) reads status.selector to count matching Pods directly,
-	// rather than via the scaled resource's own selector field.
-	selectorStr := scaleSelectorString(o.Spec.Selector)
-	return json.Marshal(map[string]any{
-		"apiVersion": "autoscaling/v1",
-		"kind":       "Scale",
-		"metadata": map[string]any{
-			"name":            name,
-			"namespace":       namespace,
-			"uid":             o.Metadata.UID,
-			"resourceVersion": o.Metadata.ResourceVersion,
-		},
-		"spec":   map[string]any{"replicas": replicas},
-		"status": map[string]any{"replicas": o.Status.Replicas, "selector": selectorStr},
-	})
-}
-
-// scaleReplicas extracts .spec.replicas from a Scale-shaped body (the PUT
-// body, or a patch already applied onto a synthesized current Scale).
-func scaleReplicas(body []byte) (int32, error) {
-	var s struct {
-		Spec struct {
-			Replicas int32 `json:"replicas"`
-		} `json:"spec"`
-	}
-	if err := json.Unmarshal(body, &s); err != nil {
-		return 0, err
-	}
-	return s.Spec.Replicas, nil
-}
-
-// setSpecReplicas returns body with spec.replicas set, preserving the rest of
-// the object. On a decode/encode error the body is returned unchanged.
-func setSpecReplicas(body json.RawMessage, replicas int32) json.RawMessage {
-	var m map[string]any
-	if err := json.Unmarshal(body, &m); err != nil || m == nil {
-		return body
-	}
-	spec, ok := m["spec"].(map[string]any)
-	if !ok || spec == nil {
-		spec = map[string]any{}
-		m["spec"] = spec
-	}
-	spec["replicas"] = replicas
-	out, err := json.Marshal(m)
-	if err != nil {
-		return body
-	}
-	return out
-}
-
-// overlayScaleWrite handles PUT/PATCH .../scale: only .spec.replicas on the
-// underlying object is settable this way (matching the real apiserver's scale
-// subresource — every other field, including the rest of .spec, is read-only
-// through this path), so unlike overlayReplace/overlayPatch there's no
-// defaultObject/identityMismatch handling here — those apply to the
-// underlying resource's own body shape, not a Scale request.
-func (h *handler) overlayScaleWrite(w http.ResponseWriter, r *http.Request, group, version, resource, namespace, name string) {
-	if !scaleSubresourceResources[resource] {
-		h.writeStatus(w, http.StatusNotFound, resource+" has no scale subresource")
-		return
-	}
-	current := h.currentObject(group, version, resource, namespace, name)
-	if current == nil {
-		h.writeStatus(w, http.StatusNotFound, "object not found: "+name)
-		return
-	}
-
-	var replicas int32
-	if r.Method == http.MethodPut {
-		body, ok := h.readObjectBody(w, r)
-		if !ok {
-			return
-		}
-		rep, err := scaleReplicas(body)
-		if err != nil {
-			h.writeStatus(w, http.StatusBadRequest, "decoding scale: "+err.Error())
-			return
-		}
-		replicas = rep
-	} else { // PATCH
-		if !supportedPatchType(r.Header.Get("Content-Type")) {
-			h.writeStatus(w, http.StatusUnsupportedMediaType,
-				"unsupported patch Content-Type "+strconv.Quote(r.Header.Get("Content-Type")))
-			return
-		}
-		patch, err := io.ReadAll(io.LimitReader(r.Body, maxWriteBytes))
-		if err != nil {
-			h.writeStatus(w, http.StatusBadRequest, "reading body: "+err.Error())
-			return
-		}
-		currentScale, err := scaleObject(namespace, name, current)
-		if err != nil {
-			h.writeStatus(w, http.StatusInternalServerError, err.Error())
-			return
-		}
-		// Scale has no listType/patchMergeKey fields to strategically merge, so
-		// treat a strategic-merge-patch request the same as a plain merge patch
-		// rather than resolving it against the underlying resource's own Kind
-		// (which has an entirely different shape than Scale).
-		patched, perr := jsonMergeOrPatch(currentScale, patch, r.Header.Get("Content-Type"))
-		if perr != nil {
-			h.writeStatus(w, http.StatusUnprocessableEntity, "applying patch: "+perr.Error())
-			return
-		}
-		rep, err := scaleReplicas(patched)
-		if err != nil {
-			h.writeStatus(w, http.StatusUnprocessableEntity, "patch did not produce a valid scale object")
-			return
-		}
-		replicas = rep
-	}
-
-	next := setSpecReplicas(current, replicas)
-	next = h.stampUpdate(next, current, group, version, resource, namespace, name, true) // scaling is a spec change
-	scale, err := scaleObject(namespace, name, next)
-	if err != nil {
-		h.writeStatus(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-	writeJSON(w, http.StatusOK, json.RawMessage(scale))
-}
-
-// jsonMergeOrPatch applies patch to current per contentType, supporting JSON
-// Patch (RFC 6902), Server-Side Apply (application/apply-patch+yaml — the
-// body is YAML, unlike every other supported type here, so it needs
-// converting before merging same as the main write path's applyPatch), and
-// treating anything else (merge-patch, strategic-merge, unspecified) as a
-// plain JSON merge patch.
-func jsonMergeOrPatch(current, patch []byte, contentType string) ([]byte, error) {
-	switch patchMediaType(contentType) {
-	case "application/json-patch+json":
-		p, err := jsonpatch.DecodePatch(patch)
-		if err != nil {
-			return nil, err
-		}
-		return p.Apply(current)
-	case "application/apply-patch+yaml":
-		j, err := yaml.YAMLToJSON(patch)
-		if err != nil {
-			return nil, err
-		}
-		return jsonpatch.MergePatch(current, j)
-	default:
-		return jsonpatch.MergePatch(current, patch)
-	}
-}
-
 func (h *handler) nowRFC3339() string {
 	if h.clock != nil {
 		return h.clock.Now().UTC().Format(time.RFC3339)
@@ -1184,57 +958,6 @@ func supportedPatchType(contentType string) bool {
 		return true
 	}
 	return false
-}
-
-// applyPatch applies a patch of the given (already-validated) content type onto
-// an existing current object (the create-on-first-apply case is handled
-// earlier, by overlayApplyCreate, before this is reached). Supports JSON merge
-// patch, JSON patch (RFC 6902), and strategic-merge patch (for built-in types,
-// via their registered schema). Server-side apply still falls back to a JSON
-// merge patch for now (real SSA field management lands in a later PR).
-func applyPatch(current, patch []byte, contentType, group, version, resource string) ([]byte, error) {
-	switch patchMediaType(contentType) {
-	case "application/json-patch+json":
-		p, err := jsonpatch.DecodePatch(patch)
-		if err != nil {
-			return nil, err
-		}
-		return p.Apply(current)
-	case "application/strategic-merge-patch+json":
-		return strategicMergePatch(current, patch, group, version, resource)
-	case "application/apply-patch+yaml":
-		// Server-side apply bodies are YAML; convert to JSON, then merge as an
-		// interim (real SSA field management lands in a later PR).
-		j, err := yaml.YAMLToJSON(patch)
-		if err != nil {
-			return nil, err
-		}
-		return jsonpatch.MergePatch(current, j)
-	default: // merge-patch
-		return jsonpatch.MergePatch(current, patch)
-	}
-}
-
-// strategicMergePatch applies a strategic-merge patch using the schema of the
-// object's built-in type: lists with a patch merge key (e.g. containers by name)
-// are merged element-wise rather than wholesale-replaced, matching the
-// kube-apiserver. Strategic merge is only defined for built-in types — the real
-// apiserver has no strategy metadata for custom resources — so a GVK that isn't
-// in the scheme falls back to a plain JSON merge patch.
-//
-// The GVK is derived from the request path's group/version/resource, not the
-// stored object body: a replayed object reconstructed from a captured LIST has no
-// apiVersion/kind (the apiserver strips TypeMeta from list items), so the body is
-// not a reliable type source.
-func strategicMergePatch(current, patch []byte, group, version, resource string) ([]byte, error) {
-	if gvk, ok := kindForResource(schema.GroupVersion{Group: group, Version: version}, resource); ok {
-		if obj, err := scheme.Scheme.New(gvk); err == nil {
-			return strategicpatch.StrategicMergePatch(current, patch, obj)
-		}
-	}
-	// Unknown/custom type: no strategy metadata, so merge like the apiserver's
-	// fallback for resources without a strategic-merge strategy.
-	return jsonpatch.MergePatch(current, patch)
 }
 
 // kindForResource resolves a plural resource name to its registered built-in Kind
@@ -1275,282 +998,6 @@ func stampTypeMeta(group, version, resource string, obj json.RawMessage) json.Ra
 		apiVersion = group + "/" + version
 	}
 	return withKind(obj, apiVersion, gvk.Kind)
-}
-
-// defaultObject applies Kubernetes API defaulting to body — e.g. an empty
-// Deployment.spec.strategy.type becomes "RollingUpdate" — matching what a real
-// apiserver does on every write. The overlay has no apiserver behind it, so
-// without this a freshly created object keeps whatever zero values the client
-// didn't set; a real controller (e.g. kube-controller-manager's deployment
-// controller) assumes defaulting already happened and errors on the zero value
-// ("unexpected deployment strategy type: \"\"") instead of treating it as
-// "use the default". Resources outside client-go's built-in scheme (CRDs) are
-// returned unchanged — there's no way to know their defaults without the CRD's
-// schema, which the overlay doesn't have.
-//
-// scheme.Scheme.Default only runs defaulters registered on the client-side
-// scheme, which — for the built-in types this project vendors
-// (k8s.io/api, not the full k8s.io/kubernetes apiserver) — is effectively
-// none of the fields real controllers care about; the actual
-// SetDefaults_Deployment-style functions live in k8s.io/kubernetes's internal
-// packages, which aren't meant to be imported as a library. applyKnownDefaults
-// hand-covers the specific, long-stable defaults our curated
-// --with-controller-manager controllers rely on instead.
-func defaultObject(gvk schema.GroupVersionKind, body json.RawMessage) json.RawMessage {
-	typed, err := scheme.Scheme.New(gvk)
-	if err != nil {
-		return body
-	}
-	if err := json.Unmarshal(body, typed); err != nil {
-		return body
-	}
-	// Marshal the client's own object back out before defaulting (rather than
-	// reusing body directly) so the merge patch computed below only reflects
-	// fields defaulting actually changed, not differences between body's
-	// exact bytes/field order and typed's. Round-tripping through the typed
-	// struct at all would normally risk silently dropping fields the
-	// vendored k8s.io/api types don't know about (e.g. a newer API field) or
-	// explicitly-sent zero values omitempty would elide — but since this
-	// undefaulted marshal is only ever diffed against the defaulted one, not
-	// returned, neither loss ends up in the result: a field defaulting
-	// doesn't touch is absent from the diff either way, and the merge patch
-	// below is applied onto the original body, preserving every field body
-	// actually had.
-	before, err := json.Marshal(typed)
-	if err != nil {
-		return body
-	}
-	scheme.Scheme.Default(typed)
-	applyKnownDefaults(typed)
-	after, err := json.Marshal(typed)
-	if err != nil {
-		return body
-	}
-	patch, err := jsonpatch.CreateMergePatch(before, after)
-	if err != nil {
-		return body
-	}
-	defaulted, err := jsonpatch.MergePatch(body, patch)
-	if err != nil {
-		return body
-	}
-	return defaulted
-}
-
-// applyKnownDefaults hand-applies the handful of long-stable Kubernetes API
-// defaults that the controllers --with-controller-manager enables (see
-// cmd/controllermanager.go) assume are already in place: a zero-valued
-// strategy/update-strategy/concurrency-policy reads as "invalid", not "use the
-// default", to those controllers. Defaulting a *Type without also defaulting
-// its matching *RollingUpdate sub-struct is not enough — real
-// kube-controller-manager code unconditionally dereferences that pointer
-// (e.g. deployment_util.go's NewRSNewReplicas reads
-// Spec.Strategy.RollingUpdate.MaxSurge), so a Type of "RollingUpdate" with a
-// nil RollingUpdate struct panics deep inside the controller instead of
-// erroring cleanly. The same function also dereferences
-// *deployment.Spec.Replicas unconditionally, so a manifest that omits
-// `replicas` (relying on the apiserver's default of 1, as e.g. Istio's charts
-// do) panics the same way — hence Deployment/StatefulSet/ReplicaSet.Replicas
-// are defaulted here too.
-
-// isIPv6Address reports whether s parses as an IPv6 address. Used to infer a
-// Service's address family from an explicit ClusterIP/ClusterIPs value
-// rather than always assuming IPv4; returns false for "" and "None" (a
-// headless Service), which correctly fall through to the IPv4 default.
-func isIPv6Address(s string) bool {
-	ip := net.ParseIP(s)
-	return ip != nil && ip.To4() == nil
-}
-
-// syntheticLoadBalancerIP deterministically derives a fake external address
-// for a LoadBalancer Service from its identity, landing in TEST-NET-3
-// (203.0.113.0/24 — reserved by RFC 5737 for documentation/example use, so it
-// can never collide with anything real). Deterministic rather than a counter
-// so the same Service gets the same address across repeated writes instead of
-// a new one every time defaultObject runs.
-func syntheticLoadBalancerIP(namespace, name string) string {
-	h := fnv.New32a()
-	_, _ = h.Write([]byte(namespace + "/" + name))
-	return fmt.Sprintf("203.0.113.%d", 1+h.Sum32()%254)
-}
-
-// syntheticClusterIP deterministically derives a fake ClusterIP for a Service
-// from its identity, landing in 10.96.0.0/12 — the conventional kubeadm
-// service-cluster-ip-range, so it looks at home next to a typical capture's
-// real ClusterIPs. Deterministic (like syntheticLoadBalancerIP) rather than a
-// counter, so the same Service keeps the same address across repeated writes.
-func syntheticClusterIP(namespace, name string) string {
-	h := fnv.New32a()
-	_, _ = h.Write([]byte("clusterip/" + namespace + "/" + name))
-	sum := h.Sum32()
-	// 10.96.0.0/12 spans 10.96.0.0-10.111.255.255: only the low 4 bits of the
-	// second octet vary (96 = 0110_0000 through 111 = 0110_1111).
-	return fmt.Sprintf("10.%d.%d.%d", 96+sum%16, (sum>>8)%256, 1+(sum>>16)%254)
-}
-
-func applyKnownDefaults(obj runtime.Object) {
-	switch o := obj.(type) {
-	case *corev1.Service:
-		if o.Spec.Type == "" {
-			o.Spec.Type = corev1.ServiceTypeClusterIP
-		}
-		if o.Spec.SessionAffinity == "" {
-			o.Spec.SessionAffinity = corev1.ServiceAffinityNone
-		}
-		// The real apiserver's IP allocator assigns every Service a ClusterIP —
-		// even LoadBalancer/NodePort ones — as soon as it's created; the overlay
-		// has no IPAM, so nothing else ever populates this. kstatus's readiness
-		// check for a LoadBalancer Service (which Helm v4's `--wait` uses via
-		// sigs.k8s.io/cli-utils) specifically requires spec.clusterIP to be
-		// non-empty — NOT status.loadBalancer.ingress — so a Service that never
-		// gets one hangs `helm install --wait` forever regardless of the
-		// synthesized external address below. A headless Service
-		// (clusterIP: "None", set explicitly by the client) and ExternalName
-		// Services are left alone — they don't get one on a real cluster either.
-		if o.Spec.Type != corev1.ServiceTypeExternalName && o.Spec.ClusterIP == "" {
-			ip := syntheticClusterIP(o.Namespace, o.Name)
-			o.Spec.ClusterIP = ip
-			if len(o.Spec.ClusterIPs) == 0 {
-				o.Spec.ClusterIPs = []string{ip}
-			}
-		}
-		if len(o.Spec.IPFamilies) == 0 {
-			// The endpoint/endpointslice controllers index IPFamilies[0]
-			// unconditionally; a real apiserver always populates this from the
-			// cluster's configured service-cluster-ip-range. Infer the primary
-			// family from the primary address only — ClusterIP if set, else
-			// ClusterIPs[0] — never by scanning every ClusterIPs entry: for a
-			// dual-stack Service with an IPv4 ClusterIP and an IPv6 secondary in
-			// ClusterIPs, scanning all entries would pick IPv6 and produce
-			// IPFamilies[0] inconsistent with the primary ClusterIP, sending the
-			// endpoint controller down the wrong address family. IPv4 remains the
-			// fallback when nothing indicates IPv6, matching every capture this
-			// project has captured so far.
-			family := corev1.IPv4Protocol
-			switch {
-			case o.Spec.ClusterIP != "" && o.Spec.ClusterIP != corev1.ClusterIPNone:
-				if isIPv6Address(o.Spec.ClusterIP) {
-					family = corev1.IPv6Protocol
-				}
-			case len(o.Spec.ClusterIPs) > 0:
-				if isIPv6Address(o.Spec.ClusterIPs[0]) {
-					family = corev1.IPv6Protocol
-				}
-			}
-			o.Spec.IPFamilies = []corev1.IPFamily{family}
-		}
-		// A real cloud-controller-manager (or an on-prem equivalent like MetalLB)
-		// eventually assigns a LoadBalancer Service an external address;
-		// --with-controller-manager's curated set deliberately excludes
-		// cloud-provider controllers (see cmd/controllermanager.go — no real cloud
-		// provider to ask), so nothing else in the overlay ever populates this
-		// field. Without it, `helm install --wait` and any other client polling
-		// for a LoadBalancer Service to become ready hangs forever. Synthesize an
-		// address deterministically from the Service's identity (rather than a
-		// counter) so repeated writes to the same Service don't reassign it.
-		if o.Spec.Type == corev1.ServiceTypeLoadBalancer && len(o.Status.LoadBalancer.Ingress) == 0 {
-			o.Status.LoadBalancer.Ingress = []corev1.LoadBalancerIngress{
-				{IP: syntheticLoadBalancerIP(o.Namespace, o.Name)},
-			}
-		}
-	case *appsv1.Deployment:
-		if o.Spec.Replicas == nil {
-			o.Spec.Replicas = ptr.To(int32(1))
-		}
-		if o.Spec.Strategy.Type == "" {
-			o.Spec.Strategy.Type = appsv1.RollingUpdateDeploymentStrategyType
-		}
-		if o.Spec.Strategy.Type == appsv1.RollingUpdateDeploymentStrategyType && o.Spec.Strategy.RollingUpdate == nil {
-			maxUnavailable := intstr.FromString("25%")
-			maxSurge := intstr.FromString("25%")
-			o.Spec.Strategy.RollingUpdate = &appsv1.RollingUpdateDeployment{
-				MaxUnavailable: &maxUnavailable,
-				MaxSurge:       &maxSurge,
-			}
-		}
-	case *appsv1.ReplicaSet:
-		if o.Spec.Replicas == nil {
-			o.Spec.Replicas = ptr.To(int32(1))
-		}
-	case *appsv1.DaemonSet:
-		if o.Spec.UpdateStrategy.Type == "" {
-			o.Spec.UpdateStrategy.Type = appsv1.RollingUpdateDaemonSetStrategyType
-		}
-		if o.Spec.UpdateStrategy.Type == appsv1.RollingUpdateDaemonSetStrategyType && o.Spec.UpdateStrategy.RollingUpdate == nil {
-			maxUnavailable := intstr.FromString("25%")
-			maxSurge := intstr.FromInt32(0)
-			o.Spec.UpdateStrategy.RollingUpdate = &appsv1.RollingUpdateDaemonSet{
-				MaxUnavailable: &maxUnavailable,
-				MaxSurge:       &maxSurge,
-			}
-		}
-		// pkg/controller/daemon/update.go's cleanupHistory dereferences
-		// *ds.Spec.RevisionHistoryLimit unconditionally (`toKeep :=
-		// int(*ds.Spec.RevisionHistoryLimit)`) — unlike the deployment
-		// controller's equivalent cleanup, which nil-checks first
-		// (HasRevisionHistoryLimit). A nil value here crashes the whole
-		// daemonset controller goroutine (client-go's crash handler logs it,
-		// then repanics — this isn't a swallowed panic like the others in this
-		// function, it takes the controller-manager process down).
-		if o.Spec.RevisionHistoryLimit == nil {
-			o.Spec.RevisionHistoryLimit = ptr.To(int32(10))
-		}
-	case *appsv1.StatefulSet:
-		if o.Spec.Replicas == nil {
-			o.Spec.Replicas = ptr.To(int32(1))
-		}
-		if o.Spec.UpdateStrategy.Type == "" {
-			o.Spec.UpdateStrategy.Type = appsv1.RollingUpdateStatefulSetStrategyType
-		}
-		// pkg/controller/statefulset/stateful_set_control.go's truncateHistory
-		// dereferences *set.Spec.RevisionHistoryLimit unconditionally
-		// (`historyLimit := int(*set.Spec.RevisionHistoryLimit)`) — same
-		// unguarded-pointer class as DaemonSet's cleanupHistory above, and
-		// likewise fatal to the whole controller-manager process, not just
-		// that one StatefulSet's reconcile.
-		if o.Spec.RevisionHistoryLimit == nil {
-			o.Spec.RevisionHistoryLimit = ptr.To(int32(10))
-		}
-		if o.Spec.PodManagementPolicy == "" {
-			o.Spec.PodManagementPolicy = appsv1.OrderedReadyPodManagement
-		}
-		if o.Spec.UpdateStrategy.Type == appsv1.RollingUpdateStatefulSetStrategyType && o.Spec.UpdateStrategy.RollingUpdate == nil {
-			partition := int32(0)
-			o.Spec.UpdateStrategy.RollingUpdate = &appsv1.RollingUpdateStatefulSetStrategy{Partition: &partition}
-		}
-	case *batchv1.Job:
-		if o.Spec.Parallelism == nil {
-			o.Spec.Parallelism = ptr.To(int32(1))
-		}
-		if o.Spec.BackoffLimit == nil {
-			o.Spec.BackoffLimit = ptr.To(int32(6))
-		}
-		if o.Spec.CompletionMode == nil {
-			o.Spec.CompletionMode = ptr.To(batchv1.NonIndexedCompletion)
-		}
-		if o.Spec.Suspend == nil {
-			o.Spec.Suspend = ptr.To(false)
-		}
-		if o.Spec.ManualSelector == nil {
-			o.Spec.ManualSelector = ptr.To(false)
-		}
-		if o.Spec.PodReplacementPolicy == nil {
-			o.Spec.PodReplacementPolicy = ptr.To(batchv1.TerminatingOrFailed)
-		}
-	case *batchv1.CronJob:
-		if o.Spec.ConcurrencyPolicy == "" {
-			o.Spec.ConcurrencyPolicy = batchv1.AllowConcurrent
-		}
-		if o.Spec.Suspend == nil {
-			o.Spec.Suspend = ptr.To(false)
-		}
-		if o.Spec.SuccessfulJobsHistoryLimit == nil {
-			o.Spec.SuccessfulJobsHistoryLimit = ptr.To(int32(3))
-		}
-		if o.Spec.FailedJobsHistoryLimit == nil {
-			o.Spec.FailedJobsHistoryLimit = ptr.To(int32(1))
-		}
-	}
 }
 
 // allowedMethods returns the Allow-header value for a write path shape, used on
